shapefil: rename files
