+++ modules/shapefile/src/dbf-loader.js
@@ -1,6 +1,6 @@
 /** @typedef {import('@loaders.gl/loader-utils').LoaderObject} LoaderObject */
 /** @typedef {import('@loaders.gl/loader-utils').WorkerLoaderObject} WorkerLoaderObject */
-import parseDBF from './lib/parsers/parse-dbf';
+import parseDBF from './lib/parsers/deprecated/parse-dbf-atomic';
 // import parseDBF from './lib/parsers/parse-dbf-state';
 
 // __VERSION__ is injected by babel-plugin-version-inline
