+++ modules/shapefile/src/shp-loader.js
@@ -1,6 +1,6 @@
 /** @typedef {import('@loaders.gl/loader-utils').LoaderObject} LoaderObject */
 /** @typedef {import('@loaders.gl/loader-utils').WorkerLoaderObject} WorkerLoaderObject */
-import parseSHP from './lib/parsers/parse-shp';
+import parseSHP from './lib/parsers/deprecated/parse-shp-atomic';
 // import parseSHP from './lib/parsers/parse-shp-state';
 
 // __VERSION__ is injected by babel-plugin-version-inline
